Delete yaourtrc
Switching AUR helper to pacaur and purging yaourt.

https://wiki.archlinux.org/index.php/AUR_helpers

Yaourt has too many security issues to be trusted with package installs from the AUR. Removing yaourtrc eliminates an unnecessary configuration file. No side effects if yaourt is not present on the system, and if yaourt is reinstalled the defeault yaourtrc will be regenerated with safer default options.